fix: move some graphql enums into their own folder to fix typeorm error for missing enums when generating migrations
