+++ src/database/enums/Marketplace.enum.ts
@@ -1,0 +1,11 @@
+import { registerEnumType } from '@nestjs/graphql';
+
+export enum Marketplace {
+  SPAACE = 'SPAACE',
+  OPENSEA = 'OPENSEA',
+  BLUR = 'BLUR',
+}
+
+registerEnumType(Marketplace, {
+  name: 'Marketplace',
+});
