+++ src/database/enums/OrderType.enum.ts
@@ -1,0 +1,12 @@
+import { registerEnumType } from '@nestjs/graphql';
+
+export enum OrderType {
+  ASK = 'ASK',
+  BID = 'BID',
+  ENGLISH_AUCTION = 'ENGLISH_AUCTION',
+  DUTCH_AUCTION = 'DUTCH_AUCTION',
+}
+
+registerEnumType(OrderType, {
+  name: 'OrderType',
+});
