+++ src/database/types/Balance.view.ts
@@ -15,11 +15,11 @@ import {
   HiddenItem,
   ItemEntity,
   LikeEntity,
-  OrderType,
   SaleEntity,
   TransferEntity,
 } from '..';
 import { utils } from '../..';
+import { OrderType } from '../enums';
 
 @ObjectType()
 @ViewEntity({
