+++ src/database/types/CollectionAttribute.view.ts
@@ -2,7 +2,8 @@ import { Field, ObjectType } from '@nestjs/graphql';
 import { ethers } from 'ethers';
 import { BaseEntity, DataSource, ViewColumn, ViewEntity } from 'typeorm';
 import { Transform } from 'class-transformer';
-import { ActiveOrderCached, ItemAttributeEntity, OrderType } from '..';
+import { ActiveOrderCached, ItemAttributeEntity } from '..';
+import { OrderType } from '../enums';
 
 @ObjectType()
 @ViewEntity({
