+++ src/database/types/DistributorReward.view.ts
@@ -2,7 +2,8 @@ import { Field, ObjectType } from '@nestjs/graphql';
 import { ethers } from 'ethers';
 import { BaseEntity, DataSource, ViewColumn, ViewEntity } from 'typeorm';
 import { Transform } from 'class-transformer';
-import { DistributorContract, DistributorRewardEntity } from '..';
+import { DistributorRewardEntity } from '..';
+import { DistributorContract } from '../enums';
 
 @ObjectType()
 @ViewEntity({
